requirements.txt
PyQt6==6.6.1
pyinstaller==6.3.0
python-dotenv==1.0.0
requests==2.31.0 